Moved keys to local.properties, and made it so randomStat on Home Screen fetches new one every 24 hours, also moved ui from ui.theme to ui.components
