+++ app/src/main/java/com/example/jetpacktest/ui/components/CircularLoadingIcon.kt
@@ -1,4 +1,4 @@
-package com.example.jetpacktest.ui.theme
+package com.example.jetpacktest.ui.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
