+++ app/src/main/java/com/example/jetpacktest/ui/components/CustomDropdown.kt
@@ -1,4 +1,4 @@
-package com.example.jetpacktest.ui.theme
+package com.example.jetpacktest.ui.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
